refactor(agents): split subagent lifecycle owners
