+++ config/max-lines-baseline.txt
@@ -507,7 +507,6 @@ src/agents/subagent-announce.format.e2e.test.ts
 src/agents/subagent-control.test.ts
 src/agents/subagent-control.ts
 src/agents/subagent-registry-lifecycle.test.ts
-src/agents/subagent-registry-lifecycle.ts
 src/agents/subagent-registry-run-manager.ts
 src/agents/subagent-registry.steer-restart.test.ts
 src/agents/subagent-registry.test.ts
